Ajout du message de débug sur discord

+++ models/Setting.py
@@ -5,7 +5,7 @@ class Setting(db.Model):
     
     id = db.Column(db.Integer, primary_key=True, autoincrement=True)
     data = db.Column(db.String(63))
-    state = db.Column(db.Integer)
+    state = db.Column(db.BigInteger)
 
     def __init__(self, data: str, state: int):
         self.data = data
